Move confirmation menu to common folder
